requirements.txt
streamlit>=1.28.0
httpx>=0.24.0
python-docx>=0.8.11
